Move package repository from JFrog to Github Packages
We've decided to move package repository (JFrog is rubbish and also costs an arm and a leg).
This change updates our connectors generation code to build connectors which by default push to GitHub, and also have a GitHub Action that packages the connector when it's tagged (as per all our other modules that aren't connectors).